Version 5.5.201.1 (cherry-pick)
Merged cb891cb06a430b35d466574cdbe7ccdecf6fe3fe
Merged 4b8f6543f497094ee33d8f9d7ae0b5a46f22c23f

Revert of [inspector] removed ArrayLengthHelperFunction specialization for empty array (patchset #2 id:20001 of https://codereview.chromium.org/2340623002/ )

Revert of [inspector] fixed all shorten-64-to-32 warnings (patchset #4 id:80001 of https://codereview.chromium.org/2332163002/ )

TBR=machenbach@chromium.org

Review URL: https://codereview.chromium.org/2344763002 .

Cr-Commit-Position: refs/heads/5.5.201@{#2}
Cr-Branched-From: 2a4699058725ebdb781d190d6d3ecdbcfd6b1343-refs/heads/master@{#39429}

+++ include/v8-inspector.h
@@ -37,14 +37,14 @@ class V8_EXPORT StringView {
  public:
   StringView() : m_is8Bit(true), m_length(0), m_characters8(nullptr) {}
 
-  StringView(const uint8_t* characters, size_t length)
+  StringView(const uint8_t* characters, unsigned length)
       : m_is8Bit(true), m_length(length), m_characters8(characters) {}
 
-  StringView(const uint16_t* characters, size_t length)
+  StringView(const uint16_t* characters, unsigned length)
       : m_is8Bit(false), m_length(length), m_characters16(characters) {}
 
   bool is8Bit() const { return m_is8Bit; }
-  size_t length() const { return m_length; }
+  unsigned length() const { return m_length; }
 
   // TODO(dgozman): add DCHECK(m_is8Bit) to accessors once platform can be used
   // here.
@@ -53,7 +53,7 @@ class V8_EXPORT StringView {
 
  private:
   bool m_is8Bit;
-  size_t m_length;
+  unsigned m_length;
   union {
     const uint8_t* m_characters8;
     const uint16_t* m_characters16;
